[#66] feat: 위도, 경도 Double로 수정

+++ collector/src/main/java/com/wherecar/collector/domain/GeoLog.java
@@ -32,10 +32,10 @@ public class GeoLog extends BaseEntity{
     private String gpsCondition;
 
     @Column(name = "latitude")
-    private Integer latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private Integer longitude;
+    private Double longitude;
 
     @Column(name = "angle")
     private Integer angle;
